% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotable_strings.R
\docType{data}
\name{annotable_strings}
\alias{annotable_strings}
\title{Annotable Strings for Unit Standardization}
\format{
A data frame with 2679 rows and 1 variable.
}
\usage{
data(annotable_strings)
}
\description{
A dataset containing commonly used strings in annotations.
}
\details{
\describe{
  \item{annotation}{A character string representing the string used as annotation.}
}
}
\keyword{datasets}
