% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_words.R
\docType{data}
\name{common_words}
\alias{common_words}
\title{Data for the common words}
\format{
A data frame with 19 rows and 9 variables.
}
\usage{
data(common_words)
}
\description{
A dataset containing data for common words.
}
\details{
\itemize{
\item Language: Contains 19 different languages.
\item Positive: Displays the word "Positive" in 19 different languages.
\item Negative: Displays the word "Negative" in 19 different languages.
\item Not_detected: Displays the phrase "Not detected" in 19 different languages.
\item High: Displays the word "High" in 19 different languages.
\item Low: Displays the word "Low" in 19 different languages.
\item Normal: Displays the word "Normal" in 19 different languages.
\item Sample: Displays the word "Sample" in 19 different languages.
\item Specimen: Displays the word "Specimen" in 19 different languages.
}
}
\keyword{datasets}
