% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_1_dummy.R
\docType{data}
\name{Function_1_dummy}
\alias{Function_1_dummy}
\title{Dummy Data for demonstrating function 1}
\format{
A data frame with 87 rows and 2 variables.
}
\usage{
data(Function_1_dummy)
}
\description{
A dataset containing dummy data for demonstrating function 1 ("clean_lab_result").
}
\details{
\describe{
  \item{raw_result}{The raw result.}
  \item{frequency}{The frequency of the raw result.}
}
}
\keyword{datasets}
