% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_2_dummy.R
\docType{data}
\name{Function_2_dummy}
\alias{Function_2_dummy}
\title{Dummy Data for demonstrating function 2}
\format{
A data frame with 86,863 rows and 5 variables.
}
\usage{
data(Function_2_dummy)
}
\description{
A dataset containing dummy data for demonstrating function 2 ("validate_lab_result").
}
\details{
\itemize{
\item patient_id: Indicates the identifier of the tested patient.
\item lab_datetime1: Indicates the date or datetime of the laboratory test.
\item loinc_code: Indicates the LOINC code of the laboratory test.
\item result_value: Indicates the quantitative result values for validation.
\item result_unit: Indicates the result units in a UCUM-valid format.
}
}
\keyword{datasets}
