% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_3_dummy.R
\docType{data}
\name{Function_3_dummy}
\alias{Function_3_dummy}
\title{Dummy Data for demonstrating function 3}
\format{
A data frame with 32 rows and 3 variables.
}
\usage{
data(Function_3_dummy)
}
\description{
A dataset containing dummy data for demonstrating function 3 ("standardize_lab_unit") containing a tiny,
intentionally messy collection of unit strings that exercise different techniques handled by the function.
}
\details{
\describe{
  \item{unit_raw}{Raw unit string to be standardized (character)}
  \item{n_records}{Optional frequency used to test the \code{n_records} argument (integer)}
  \item{note}{Human-readable tag for the test case (character)}
}
}
\keyword{datasets}
