% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_4_dummy.R
\docType{data}
\name{Function_4_dummy}
\alias{Function_4_dummy}
\title{Dummy Data for demonstrating function 4}
\format{
A data frame with 48 rows and 3 variables.
}
\usage{
data(Function_4_dummy)
}
\description{
A dataset containing dummy data for demonstrating function 4 ("harmonize_lab_unit") including
different success (harmonized) and failure (not_harmonized) cases handled by the function.
}
\details{
\describe{
  \item{loinc_code}{LOINC code of the laboratory test.}
  \item{result_value}{Quantitative result value for validation.}
  \item{result_unit}{Result unit in UCUM-compliant format.}
}
}
\keyword{datasets}
