% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_rules.R
\docType{data}
\name{logic_rules}
\alias{logic_rules}
\title{Data for the logic rules}
\format{
A data frame with 18 rows and 4 variables.
}
\usage{
data(logic_rules)
}
\description{
A dataset containing data for the logic rules.
}
\details{
\describe{
  \item{rule_id}{Identifier for the logic rule.}
  \item{rule_index}{The sequence index of the rule.}
  \item{rule_part}{The textual content of the rule part.}
  \item{rule_part_type}{The type/category of the rule part (e.g., operator, term, value).}
}
}
\keyword{datasets}
