% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseUnit_function_V4.R
\name{parseUnit}
\alias{parseUnit}
\title{Parse a UCUM unit string}
\usage{
parseUnit(uStr, units_df)
}
\description{
INTERNAL helper.  Converts a raw unit string to a row compatible with
`units_df`.
}
\keyword{internal}
