% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportable_interval.R
\docType{data}
\name{reportable_interval}
\alias{reportable_interval}
\title{Data for the reportable interval}
\format{
A data frame with 493 rows and 4 variables.
}
\usage{
data(reportable_interval)
}
\description{
A dataset containing data for the reportable interval.
}
\details{
\itemize{
\item interval_loinc_code: The interval of the LOINC code.
\item UCUM_unit: The UCUM unit.
\item low_reportable_limit: The lower reportable limit.
\item high_reportable_limit: The higher reportable limit.
}
}
\keyword{datasets}
