% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_words.R
\docType{data}
\name{common_words}
\alias{common_words}
\title{Data for the common words}
\format{
A data frame with 19 rows and 9 variables.
}
\usage{
data(common_words)
}
\description{
A dataset containing data for common words.
}
\details{
\describe{
  \item{Language}{The name of the language.}
  \item{Positive}{Translation of the word "Positive".}
  \item{Negative}{Translation of the word "Negative".}
  \item{Not_detected}{Translation of the phrase "Not detected".}
  \item{High}{Translation of the word "High".}
  \item{Low}{Translation of the word "Low".}
  \item{Normal}{Translation of the word "Normal".}
  \item{Sample}{Translation of the word "Sample".}
  \item{Specimen}{Translation of the word "Specimen".}
}
}
\keyword{datasets}
