% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_2_dummy.R
\docType{data}
\name{Function_2_dummy}
\alias{Function_2_dummy}
\title{Dummy Data for demonstrating function 2}
\format{
A data frame with 86,863 rows and 5 variables.
}
\usage{
data(Function_2_dummy)
}
\description{
A dataset containing dummy data for demonstrating function 2 ("validate_lab_result").
}
\details{
\describe{
  \item{patient_id}{Identifier of the tested patient.}
  \item{lab_datetime1}{Date or datetime of the laboratory test.}
  \item{loinc_code}{LOINC code of the laboratory test.}
  \item{result_value}{Quantitative result value for validation.}
  \item{result_unit}{Result unit in UCUM-compliant format.}
}
}
\keyword{datasets}
