% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_rules.R
\docType{data}
\name{logic_rules}
\alias{logic_rules}
\title{Data for the logic rules}
\format{
A data frame with 18 rows and 4 variables.
}
\usage{
data(logic_rules)
}
\description{
A dataset containing data for the logic rules.
}
\details{
\itemize{
\item rule_id: The rule ID.
\item rule_index: The rule index.
\item rule_part: The rule part.
\item rule_part_type: The rule part type.
}
}
\keyword{datasets}
