% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportable_interval.R
\docType{data}
\name{reportable_interval}
\alias{reportable_interval}
\title{Data for the reportable interval}
\format{
A data frame with 493 rows and 4 variables.
}
\usage{
data(reportable_interval)
}
\description{
A dataset containing data for the reportable interval.
}
\details{
\describe{
  \item{interval_loinc_code}{The LOINC code to which the reportable interval applies.}
  \item{UCUM_unit}{The UCUM-compliant unit for the laboratory measurement.}
  \item{low_reportable_limit}{The lower limit of the reportable range.}
  \item{high_reportable_limit}{The upper limit of the reportable range.}
}
}
\keyword{datasets}
